Revert "preparing for writing an improved makeworld script" et al.
This change reverts commits c2231b4f..1c80046c. It is not a good idea to have
two separate scripts in this repository that are meant to do the exact same
thing. That situation is bound to confuse our users. If there are improvements
to be made to `makeworld`, then they should be merged into that file so that
everyone benefits from them. Significant changes should go through a pull
request to give everyone who has an interest in this script the chance to
review them. The development of those improvements up to that point should take
place in a forked repository or in a branch.