Remove Chart and its dependencies.
The new version of Chart (1.0) depends on lens, which in turn has an upper
limit on its dependency on split (<0.2.2) which is lower than what is already
in the repo (0.2.2).

Reason for removal rather than waiting out developers of lens:
The current version of lens was released late August 2013 and still depends on
split <0.2.2 despite split 0.2.2 being released already in April 2013.  This
suggests that adding lens at this point risks keeping split back for a long
time.

Signed-off-by: Magnus Therning <magnus@therning.org>