Add support to pass in uid and gid to build scripts.
The building is performed as root inside the docker image, so an explicit
`chmod` step is required to ensure the files have good ownership.  Previously
the uid and gid where hard coded, now they can be controlled when invoking
`docker-bild` and they are passed on to the build script running inside
docker, `helpers/buildpkgs`.

Signed-off-by: Magnus Therning <magnus.therning@evidente.se>