Fix up `hspec`.
Add `hspec-discover` and add it as a dependency of `hspec`.

The upstream solution to ensure that the tool `hspec-discover` is installed
as a dependency of `hspec` is to create a dummy library and then add it as a
dependency.  Is there really no better way to ensure that the tool is
installed?  Some sort of tool dependency in Cabal?

The solution here is rather roundabout :(  But it works.

Signed-off-by: Magnus Therning <magnus@therning.org>